Rework a little bit of a8a1994d89c8d24f86b4e6920a3901c17aeed922 so that we don't apply a `search_path` to every function by default.
Now there's a new `#[search_path(...)]` attribute macro you can apply to pg_extern/operator/test functions to set the search path.

There's some bit of per-execution overhead for functions with a custom search_path, which @a-mckinley discovered.

This resolves issue #58

A drop/create of your extension will be necessary in order for the new function declarations to be created.