`PgBox` rewrite (#274)
Rewrites the internals of `PgBox` such that it's now the size of a pointer, instead of twice the size of pointer.

It introduces some zero-sized types (ZSTs) to indicate if the pointer owned by a PgBox instance was allocated by Rust or by Postgres, which might cause compilation errors for existing users. Its default understanding is `AllocatedByPostgres`.