Introduce `pgx::slice::PallocSlice` (#882)
* Introduce `pgx::slice::PallocSlice`

This is an abstraction designed to simplify handling of pallocs
when those pallocs are also slices. Its semantics differ from PgBox
a little bit on a few points so it isn't the same type.

* Refactor `datum::Array` with `PallocSlice`

This removes the "double pointer" scenarios the impl of Array
had to constantly be wary of and guard against, for both palloc
and the detoasting concern. The Drop remains slightly touchy
and a needless Option remains, but that can be fixed.

* Reuse get_unchecked in PallocSlice::get

* Remove lifetime from PallocSlice

Note the usage limitation in the comments for the type,
as we currently don't accurately model this problem.