Fix a bug trying to return `Result<pg_sys::Oid>` (#976)
HACK for `Result<T, E>` -- we replace the `E` with an underscore

Specifically, this is the cases where we need to use the `map_source_only!()` macro to create a Rust type to SQL type mapping.  As of this writing that's only for the `pg_sys::Oid` type, but we'll do our best to generalize this.

This also teaches `map_source_only!()` to fabricate `Result` and `Result<Option>` wrappers for the Rust type being mapped.