Proper error handling for SPI (#977)
- the getter methods on PgTupleTable and SpiHeapTupleData go through `try_from_datum()`, and 
- many methods like `Spi::run`, `SpiClient::update`, `SpiClient::select` now return a `Result<R, spi::Error>`.
- cleanups to the `spi::Error` variants
- make a few internal pointers we track `NonNull`
- added some comments here and there, and updated all the tests/examples accordingly.

Prior to this PR `spi.rs` was silently transforming non SpiOK returns into a panic!, and now those are pushed forward to the user as a `spi::SpiError` (either directly or indirectly as a variant of `spi::Error`).  If a Postgres SPI function returns an error that probably means its recoverable, so we really need to pass them back to the user.  This includes methods like `Spi::run()`, `SpiClient::update()`, etc.

- `Spi::connect()` now returns `-> R` instead of `spi::Result<R>`, and contrary to the above statement, it will `panic!()` if `pg_sys::SPI_connect()` fails
- everything in `spi.rs` now returns `spi::Error` (no more returning `spi::SpiError`)
- `lib.rs` now only exports `spi::Spi` and the `crate::spi` module itself so as not to pollute the user's namespace
- rename `SpiOk` and `SpiError` to `SpiOkCodes` and `SpiErrorCodes` to indicate that they ought not be used as `Result:Err()`, especially within `spi.rs`.  We prefer the more general `spi::Error` enum.

Thanks to @EdMcBane & @yrashk for the review and testing of these changes!