Extend Spi's understanding of "readonly" to transaction (#992)
Mucking with the pushing a new Snapshot if we *think* we've modified the database isn't correct.  It can cause problems elsewhere within Postgres where Postgres expects the Portal's snapshot to be the same as when the Portal was created.

Track if `Spi` is in `MUTABLE_MODE` or not for the duration of the transaction.