`pgx` shouldn't use `register_xact_callback` to help track a transaction's mutability.
This can cause problems if for some reason the pgx extension is unloaded from a running postgres backend as Postgres will still have a pointer to the transaction hook function that has now been unloaded. (#1104)

Instead, we can use the current `TransactionId` to determine if the transaction has been mutated, which still allows us (more correctly, even), to ensure Spi sets the `read_only` flag appropriately when executing a statement.

This also gets rid of a pgx-specific static.