Use zero-copy Arrays in `--release` (#1116)
Implementing truly "zero-copy" arrays requires many internal changes.

## `Toast<T: Toasty>`
Toast and Toasty lift detoasting and pfreeing data into the type level.
This is used in `datum::Array` to remove its Drop and allow deconstruction.
It also makes that code "automatically correct" except for the usual suspects:
- ManuallyDrop
- `mem::forget`
- unsafe

## Array-walking code

An implementation of "walking" Postgres arrays in Rust is added to the library
for internal usage. A variant may be exposed in the future if such is desired.

Postgres arrays are a kind of varlena that internally has an additional fairly
complex suite of layout rules governing how the types are laid out within it.
This means Postgres has many small exceptions in its layout code for arrays.
Fortunately, most of these exceptions are to make the array layouts simpler.
Unfortunately, it also means not all Postgres arrays have fixed element sizes.
Indeed, some elements in a Postgres array are actually "zero size types":
they occupy 0 bytes inline, and 1 bit (`0b0 << ?`) in the array's nullbitmap.

Thus indexing into Postgres arrays is not always a single offset computation,
and Postgres must manually "walk" the array i steps to find the ith index.

## Debug assertions

The previous version of the code uses `pg_sys::deconstruct_array` instead,
yielding arrays of precalculated pointers instead, but this needs allocation
and significant overhead waiting for Postgres to calculate all the offsets.
This version of the code is preserved in an altered form for debug assertions.
However, even this is a bit much for tests, so it is slated for removal soon.
These guard rails proved sufficiently useful during development, however, that
removing them before users have a chance to run them against their own tests
seems like it would be premature. More lightweight assertions should remain,
and some of these assertions can be converted into lighter-weight forms.