fix issue #1141:  immediately report FATAL and PANIC (#1142)
errors reported at the FATAL or PANIC levels should immediately be reported
to Postgres since they ultimately abort the transaction.

ERRORs continue to be converted into panics, and everything else is also
immediately reported to Postgres but since they only emit messages,
there's no change to flow control.

As a drive-by, I also added a few functions to the "bad_ideas" example to interactively test this change.  It's not really feasible for the test suite to somehow test that a test causes the backend, or the entire cluster, to crash, which is what PANIC and FATAL do.