Add proptest support (#1258)
Our proptest support requires `PgTestRunner`, which shims `TestRunner`
in what will be increasingly custom ways, in order to support running
inside `#[pg_test]`.

Most of this PR is documentation on the expected patterns and what
actually still works, as a result. Additional work on adding value trees
for complex-to-construct types will come later.

What I thought would also be an "easy sample test" demonstrated we
shouldn't have a `From<pg_sys::DateADT>` (an alias for i32) impl for
Date. Thus this also breaks public API by removing `From<i32> for Date`,
downgrading it to a `TryFrom` that respects the boundaries found by the
test's shrinking to simplified case. `Date::saturating_from_raw` is
added for accepting the inner type and saturating it to `-infinity` or
`infinity` if it is out-of-range for a Date. This is necessary in order
to make the "sample" property test pass.