Assume commutation when deriving PostgresEq (#1261)
This fixes https://github.com/pksunkara/pgx_ulid/pull/16 without
requiring additional work on the part of the programmer. However, there
is a design decision here: do we require explicit annotation before
assuming commutative equality or not?

Here I chose to demand that the equality implementation adheres to
Rust's requirements for "total" equality. However, as the SQL we emit is
used deep inside Postgres for various forms of algebraic reasoning,
including, as I understand it, index maintenance, I can't promise that
we don't wind up running afoul of Postgres using this data in a way that
will incur UB if the implementation is actively malicious, as can be
done with implementations of the following kind:

```rust
impl PartialEq for AlwaysRandom {
    fn eq(&self, other: &Self) -> bool {
        rand::random()
    }
}

impl Eq for AlwaysRandom { }
```

Of course, this also violates PARALLEL SAFE and IMMUTABLE.
This problem also was always true, as we emitted NEGATOR clauses,
which means even this was probably already broken:

```rust
impl PartialEq for AlwaysWrong {
    fn eq(&self, other: &Self) -> bool {
        false
    }
}

impl Eq for AlwaysWrong {}
```

This is less likely to cause issues in more realistic code samples.