Make `PostgresHash` also need `Eq` (#1264)
Latest in series of fixing the derive macros to require the correct
bounds. This time, basically, "The hash operators are used for hashmappy
things, so they should have the same bounds as hashmaps."

Moves away from the briefly-adopted `#[doc(hidden)]` trait pattern to
instead using a "trivial bound" for these. It didn't actually occur to
me, first go-around, that such was correct in Rust and would emit a
similar, reasonably-nice error.