Introduce pure-Rust replacement for PgList (#1277)
I'm allowing the original PgList type to continue to exist for code that
might still be relying on PgList for a bit longer and because this way
we can have a deprecation period for the original type.

The high level overview:
- `*mut pg_sys::List` is used like a peculiar cousin of a
`Vec<pg_sys::ListCell>` since Postgres 13. Before that it was a true
linked list. This is wrapped by `enum pgrx::list::List<T>` which is
really a funny-looking `Option<ListHead<T>>`.
- ...`ListHead<T>` is itself a funny-looking `NonNull<pg_sys::List>`.
- These allow obtaining `{&,&mut} [ListCell<T>]`, a slice of
strongly-typed `pg_sys::ListCell`s.
- `unsafe trait Enlist` describes a union variant of `pg_sys::ListCell`.
- All ways to poke inside a `ListCell<T>` are bounded by `T: Enlist`.
- All ways to get a `List<T>` require `unsafe` assertions about the
List's validity for various things... unless it's a null pointer, oddly
enough. That is, `List::Nil` is publicly constructible quite on purpose,
and otherwise you need to construct a `ListHead<T>` first (only via an
`unsafe fn`), or you can't build the other variant of `List<T>`.

~~This is currently bounded by Postgres version~~ Fixed that.

In order to obtain some coherent ergonomics for `pg_sys::NodeTag`, it is
now a Rust enum. This is kosher because **all node tag values can change
when you add even one variant**, so it is not quite as unstable as a
*true* Rust enum, but it *is* incoherent if you aren't built directly
against the EXACT same version of Postgres. They are NOT stable from
version to version, except, fortunately, for the zeroed variant
(`NodeTag::T_Invalid`).