Consistently order generated impls (#1314)
Using BTreeSet instead of HashSet allows ceasing to permeate
`is_for_release` through bindgen, into parts which don't need to know
about that detail. This removes a source of inconsistency in our build
artifacts, and also avoids having to do an additional reallocate and
sort step, by simply using a type that is always going to be sorted.

Local testing shows no negative impacts on compile time, and a mild
positive, but within noise variance... this is only a couple seconds on
a compile that takes almost a minute, even on a powerful machine.