Touch up Display for PgNode (#1313)
This wasn't using the unambiguous root path addressing.

Avoiding `write!` is often faster for fmt impls, simply because the
macro tends to generate subpar code. This now dispatches to the String
fmt impl, which respects alignment and padding requests, so it's hard to
say if it will be faster, but in any case it's more correct.