Add trusted option in the control file (#1397)
`trusted` is supported in postgresql 13 and above. Extensions can add it
to the control file and also want to support postgresql 12 and below.
This PR removes the trusted option in the control file when installing
the extension. The default value of `trusted` is false.

Options in the control file:
https://www.postgresql.org/docs/13/extend-extensions.html#id-1.8.3.20.11

The following is rejected as it is invalid:
```
superuser = false
trusted = true
```