Make sql-entity-graph clippy-clean (#1420)
Resolve the final batch of clippy warnings in the sql-entity-graph crate
and deny further appearances of those warnings. This is to prevent
visibly half-baked code from being merged to this extremely
functionally-sensitive crate, since the failure to do so made the code
so unreadable that no one touched it for a year.