Simplify void return handling (#1422)
Aside from imposing a few stylistic choices on pgrx-sql-entity-graph,
FunctionMetadata has its second generic removed after some consulting
with type system experts who expressed it is unlikely to be necessary.
Indeed, everything seems to be unaffected by this.

In doing this, it reveals that FunctionMetadataEntity is only using
Option to handle a single return type: `()`. Removing this allows
simplifying the handling for function pointers that return `-> ()`... of
which you might remember there are a lot, given that is any function
without an explicit other return type.