Demand lifetimes for `pgrx::composite_type!` (#1462)
The `pgrx::composite_type!` macro currently assigns the `'static`
lifetime if you do not give one. This is a decision of dubious
soundness. Instead, demand a lifetime, or assign the anonymous lifetime.
Both cases match the behavior of attempting to elide lifetimes in your
usual Rust code, though with slightly worse error messages in some
cases. This may require people to write `'static` in more places (and
that is what the typical error suggests), but it allows noticing and
fixing problems that currently are not visible, easy to understand, or
easy to check.

This is the latest in this series:
- https://github.com/pgcentralfoundation/pgrx/issues/1383
- https://github.com/pgcentralfoundation/pgrx/pull/1457
- https://github.com/pgcentralfoundation/pgrx/pull/1461