Use `composite_type!` lifetimes in pgrx-sql-entity-graph (#1463)
I was working on fixing the auto-staticizing lifetime issues and noticed
this likely would be landable on its own. This moves out `struct
CompositeTypeMacro` into its own file and starts to actually modularize
the code around it a little, to start actually preserving the lifetimes
in their usage during macro expansion. Because we need to also generate
wrapping types, it doesn't count as much of a cleanup on its own, but
that's why I'm landing it separately.