Filter `all` to both config && Cargo .tomls (#1487)
Our support for a given Postgres version, for an extension crate, is
informed by two things:
- The setting to find `pg_config`s, defaulting to
`"${PGRX_HOME}/config.toml"`
- The crate's Cargo.toml for a given pgrx `"pg{major}"` feature.

Interpret `cargo-pgrx {start,stop,test} all` to act only on versions for
which both of these are present. We will simply filter silently without
note because there are some nuances for when we should (not) emit errors
that I am not entirely settled on how to decide, but ultimately this
allows e.g. an extension like PL/Rust that never supported the same
version minimum as pgrx to still use `cargo pgrx test all`, and causes
less confusing errors when `config.toml` includes out-of-date versions
unless those are also in Cargo.toml.

Note this doesn't prevent us from emitting errors: instead the error
moves to when building the pgrx library. Currently it is first caught by
cargo when the extension's Cargo.toml doesn't include the features to
for the request from cargo-pgrx, but this is a hidden behavior of
cargo-pgrx. The failed build of the library is easier to diagnose.