Fixup docs for cargo-pgrx (#1491)
Remove mentions of commands that do not exist, versions that we do not
support, and contradictory or confusing statements that, at best, rely
on remembering what overrides what. The last are in most cases edited
into simply treating variables as variables, or simply deleted to avoid
disagreeing with the detailed explanation that naturally arises with
further reading.