Document we compile Postgres with assertions (#1513)
From the conversation with @rjuju in pgcentralfoundation/pgrx#1511. We
did not document how we compile Postgres, which results in some
confusion when people expect to need to have to go to additional trouble
to test against a Postgres with debug symbols and assertions. Let's
write that down: we always build a debuggable Postgres with asserts.

As a matter of spring cleaning, this includes some tangential fixups
that crossed my eyes as I rebuilt the docs, and some clarification
of the rationales for cargo-pgrx's behavior.