Bounds-checking debug assertions for array access (#1514)
This introduces several `debug_assert!()` calls around any attempt to
dereference an element in a Postgres array, in order to catch memory
bugs. It was written to resolve pgcentralfoundation/pgrx#1195.

If RawArray's `end_ptr()` ever pointed at something other than one byte
past the end of the array's buffer, this might fail to catch memory
bugs. However, it should be pointing to the end of the ArrayType's
varlena, which includes the ArrayType header, its null bitmap, and its
data, so this should work.