Add "installation" section to docs (#1569)
This addresses a few questions that keep coming up.

Resolves https://github.com/pgcentralfoundation/pgrx/issues/1538

Also quite a bit of rearrangement was done while I was thinking about
this, because our documentation is a bit hard to navigate even if you
build it, as it relies on the summary bar to the left instead of being
navigable via in-doc hyperlinks. This makes it easy to forget where you
are supposed to be going if you are referencing the Markdown itself.

I focused on trying to make it so the SUMMARY.md can be repeated...
copy-pasted, really... into each section for the relevant parts.