Pipe forward spans by formatting idents (#1665)
Apply a simple change to -sql-entity-graph and -macros:

Use `format_ident!` for every case we format... an identifier... and we
don't change the identifier, and reuse the existing span or worse, drop
the span on the floor by calling `Span::call_site`. This will, in
general, improve the error-reporting of spans so that errors refer back
to the original input that provoked the expansion.

It is possible we should by relocating synthesized spans, see
https://github.com/rust-lang/rust/issues/124145 for more details. This
change is still preferred because it makes it easier to refactor these
again.