Use UnsafeCell for PgLwLock and PgAtomic (#1696)
As explained in detail in #1223, due to the Postgres restart-on-crash
logic, `PgAtomic<T>::attach` and `PgLwLock<T>::attach` would be called
again after a crash. Using OnceCell prevented the restarted extension
from attaching new PgLwLocks, so Postgres failed to restart after a
crash. This addresses that by using UnsafeCell instead of OnceCell and
marking `attach` as unsafe.