Implement `RetAbi` and `BoxRet` (#1701)
This completely reworks the way we handle returning types from
functions, so that we no longer have to rely on a macro expansion
behavior that has to already know the types at expansion time (and thus
has to parse them somehow, which it cannot realistically do because type
aliases exist). Instead, we simply expand into code that asks the types
themselves to modify the FunctionCallInfo appropriately and then return
a raw Datum to Postgres.

This breaks support for certain returns because it was difficult to do
this and also support arbitrary nesting, because Postgres does not
support arbitrary nesting. For instance, you can no longer return:
- `SetOfIterator<'a, Result<T, E>>`
- `TableIterator<'a, (Result<T, E>, Result<U, D>)>`
- `Option<TableIterator<'a, Tuple>>`

It's expected that this will improve in the near-ish future.

This also breaks returning values from `#[pg_extern]` functions that
were relying on `IntoDatum` implementations being enough. It is not
expected this will improve, for the reasons described on the
documentation of the new traits, which can be summarized as "`IntoDatum`
should never have been used for that bound". This change blocks off
several latent correctness problems from affecting pgrx going forward.

Fixes https://github.com/pgcentralfoundation/pgrx/issues/1484