Refactor BoxRet to be easier to impl before showing it (#1776)
People need to implement some of the code in this module, but I want to
make sure it is easier to actually do so. To that end, refactor BoxRet
to have a more user-friendly interface, ish, and then expose it.

While doing so, adjust some safety requirements here and there.