Add bindings for jit and resowner headers (#1790)
These headers are required when writing a JIT provider for PostgreSQL.
Most things in jit.h are required, and from resowner_private.h the
ResourceOwner*JIT() functions are necessary. Additionally jit.h contains
the struct used when measuring how much time is spent doing just-in-time
compilation which is useful when working with EXPLAIN ANALYZE outputs.

The resowner_private.h header will be removed in PostgreSQL 17,
and replaced entirely by new facilities in resowner.h so
to make migration easy that header is also included.