Use `petgraph::algo::tarjan_scc` to order the schema graph (#1867)
This seems to give a consistent ordering of the schema through
consecutive runs of `cargo pgrx schema`.

It also allows `diff` to produce nice diffs when the schema changes
compared to a previous version.

I've tested this ordering with the `pg_search` extension (which is quite
large) and it still generates a perfectly usable schema. The same one
from run to run, even!

It also adds comment markers around each connected block of sql objects,
just to help better visualize the connectedness of the schema.