Add `BorrowDatum` for unsizing borrows of datums (#1891)
Add a trait that allows borrowing Datums from various places, like
arguments and arrays, instead of always taking them by-value. This is
most important for unsizing borrows, as the sized kind are a simple
`.cast()` away.

The motivating rationale is to enable new APIs in the near future like
FlatArray and Text. It also brings in some more uniformity because of
the blanket-impl of ArgAbi for `&T`. This means that various types that
_are_ pass-by-reference may now actually be taken by reference from
their Datum, instead of copying them into the Rust function! This is
more important for unsized types, but for various reasons, right now it
only takes over the ArgAbi of one unsized type: CStr.